+++ GUI/studygraphicswidget.cpp
@@ -168,7 +168,7 @@ template<typename T> void load_rgb_image2(
 	{
 		const double tmp_max =
 			(bits_allocated > 0 && bits_stored > 0 && bits_stored < bits_allocated)
-			? pow(2, bits_stored) - 1
+			? pow(2.0, static_cast<double>(bits_stored)) - 1.0
 			: static_cast<double>(USHRT_MAX);
 		try
 		{
@@ -348,7 +348,7 @@ template<typename T> void load_rgba_image2(
 		{
 			const double tmp_max =
 				(bits_allocated > 0 && bits_stored > 0 && bits_stored < bits_allocated)
-				? pow(2, bits_stored) - 1
+				? pow(2.0, static_cast<double>(bits_stored)) - 1.0
 				: static_cast<double>(USHRT_MAX);
 			try
 			{
@@ -414,7 +414,7 @@ template<typename T> void load_rgba_image2(
 		{
 			const double tmp_max =
 				(bits_allocated > 0 && bits_stored > 0 && bits_stored < bits_allocated)
-				? pow(2, bits_stored) - 1
+				? pow(2.0, static_cast<double>(bits_stored)) - 1.0
 				: static_cast<double>(USHRT_MAX);
 			try
 			{
